+++ Core.lua
@@ -17,8 +17,26 @@ Core.Libs.LibRu = LibRu;
 -- register default slash command
 Core:RegisterSlashCommand("/bettertransmog", {
     ["default"] = function(msg, editbox)
-        Core:DebugLog("BetterTransmog default command executed: " .. tostring(msg) .. " | Editbox: " .. tostring(editbox))
-    end
+        ---@type BetterTransmog.Modules.Settings
+        local settingsModule = Core:GetModule("Settings");
+        if settingsModule then
+            settingsModule:OpenSettingsFrame();
+        end
+    end,
+    ["settings"] = function(msg, editbox)
+        ---@type BetterTransmog.Modules.Settings
+        local settingsModule = Core:GetModule("Settings");
+        if settingsModule then
+            settingsModule:OpenSettingsFrame();
+        end
+    end,
+    ["changelog"] = function(msg, editbox)
+        ---@type BetterTransmog.Modules.ChangeLog
+        local changeLogModule = Core:GetModule("ChangeLog");
+        if changeLogModule then
+            changeLogModule:ShowChangeLog();
+        end
+    end,
 })
 
 Core:RegisterSlashCommand("/rl", function (msg, editbox)