+++ CHANGELOG.md
@@ -1,14 +1,33 @@
 # Version 2.0.4
+Quite a major update containing many requested features and improvements! As usual, thank you for your feedback and support, if you have any, don't hesitate to open a issue on the github page!
+
 
 ## Added BetterTransmog Minimap button
-Added a better transmog minimap button that lets you open the Outfit Swap mode or the settings
- 
+Added a better transmog minimap button that lets:
+- \[Left-Click\] Open Outfit swapping panel 
+- \[Right-Click\] Open BetterTransmog Settings Panel
+
 ![Minimap Button Screenshot](Assets/Screenshots/ChangeLog_MinimapButtonExample.png)
 
+## Added new /slashcommands! 
+The slashcommand system is still a WIP, i will add aliases to command as i improve it in a later update!
+Here are the new /slashcommands:
+- `/bettertransmog` or `/bettertransmog settings` (opens settings panel)
+- `/bettertransmog outfits` (Opens the outfit swapping panel)
+- `/bettertransmog changelog` (Opens the changelog panel)
+
+Expect improvements to slashcommands in the next update so you can alias `/bettertransmog` to `/bt`
+
+## Improved Changelog Frame and System
+Made the changelog frame more user-friendly with better readability and smoother interactions, so you can enjoy reading about updates without any hassle!
+
+## Codebase upgrades
+I won't bore you with the technical stuff, but under the hood, things are changing and improving! and if you are interested, you can always have a peek at the GitHub repo <3
 
 # Version 2.0.3
 
 ## Fixed a bug causing a lua error in the Resizing Module
+Yep, gotta make sure we keep it clean and lua error free! 07
 
 # Version 2.0.2
 
@@ -20,3 +39,5 @@ Introduced a new ChangeLog module that displays version updates to users upon ad
 
 ## Added a /rl command. |r
 Added a shorthand way to /reload the ui with /rl (Request by F0ki & Jimbo) |r
+
+